janji_temu_pengguna/requirements.txt
Flask==2.3.3python app.py

mysql-connector-python==8.0.33